moving datastructures to new package datastructures
