+++ Library-SMTLIB/src/de/uni_freiburg/informatik/ultimate/logic/FormulaUnLet.java
@@ -22,7 +22,7 @@ import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Map;
 
-import de.uni_freiburg.informatik.ultimate.util.ScopedHashMap;
+import de.uni_freiburg.informatik.ultimate.util.datastructures.ScopedHashMap;
 
 /**
  * This class removes all let terms from the formula.  The result
