+++ Library-SMTLIB/src/de/uni_freiburg/informatik/ultimate/logic/Theory.java
@@ -28,8 +28,8 @@ import java.util.LinkedHashSet;
 import java.util.Map;
 
 import de.uni_freiburg.informatik.ultimate.util.HashUtils;
-import de.uni_freiburg.informatik.ultimate.util.ScopedHashMap;
-import de.uni_freiburg.informatik.ultimate.util.UnifyHash;
+import de.uni_freiburg.informatik.ultimate.util.datastructures.ScopedHashMap;
+import de.uni_freiburg.informatik.ultimate.util.datastructures.UnifyHash;
 
 
 /**
