+++ Library-SMTLIB/src/de/uni_freiburg/informatik/ultimate/util/datastructures/ScopedHashMap.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with SMTInterpol.  If not, see <http://www.gnu.org/licenses/>.
  */
-package de.uni_freiburg.informatik.ultimate.util;
+package de.uni_freiburg.informatik.ultimate.util.datastructures;
 
 import java.util.AbstractMap;
 import java.util.AbstractSet;
@@ -24,6 +24,8 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
+
+import de.uni_freiburg.informatik.ultimate.util.ScopeUtils;
 
 
 /**
