+++ Library-SMTLIB/src/de/uni_freiburg/informatik/ultimate/util/datastructures/ScopedHashSet.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with SMTInterpol.  If not, see <http://www.gnu.org/licenses/>.
  */
-package de.uni_freiburg.informatik.ultimate.util;
+package de.uni_freiburg.informatik.ultimate.util.datastructures;
 
 import java.util.Arrays;
 import java.util.Collection;
