+++ SMTInterpol/src/de/uni_freiburg/informatik/ultimate/smtinterpol/convert/ArithDelay.java
@@ -30,7 +30,7 @@ import de.uni_freiburg.informatik.ultimate.logic.Sort;
 import de.uni_freiburg.informatik.ultimate.logic.Term;
 import de.uni_freiburg.informatik.ultimate.logic.TermTransformer;
 import de.uni_freiburg.informatik.ultimate.logic.Theory;
-import de.uni_freiburg.informatik.ultimate.util.ScopedHashMap;
+import de.uni_freiburg.informatik.ultimate.util.datastructures.ScopedHashMap;
 
 public class ArithDelay extends InternTermTransformer {
 	private final ScopedHashMap<Rational, Term> mArithConsts =
