+++ SMTInterpol/src/de/uni_freiburg/informatik/ultimate/smtinterpol/smtlib/smtlib.cup
@@ -20,7 +20,7 @@
 
 package de.uni_freiburg.informatik.ultimate.smtinterpol.smtlib;
 import de.uni_freiburg.informatik.ultimate.logic.*;
-import de.uni_freiburg.informatik.ultimate.util.ScopedHashMap;
+import de.uni_freiburg.informatik.ultimate.util.datastructures.ScopedHashMap;
 import com.github.jhoenicke.javacup.runtime.Symbol;
 import de.uni_freiburg.informatik.ultimate.smtinterpol.util.MySymbolFactory;
