+++ SMTInterpol/src/de/uni_freiburg/informatik/ultimate/smtinterpol/smtlib2/smtlib.cup
@@ -20,7 +20,7 @@
 
 package de.uni_freiburg.informatik.ultimate.smtinterpol.smtlib2;
 import de.uni_freiburg.informatik.ultimate.logic.*;
-import de.uni_freiburg.informatik.ultimate.util.ScopedHashMap;
+import de.uni_freiburg.informatik.ultimate.util.datastructures.ScopedHashMap;
 import java.math.BigInteger;
 import java.math.BigDecimal;
 import java.util.Arrays;
