+++ SMTInterpol/src/de/uni_freiburg/informatik/ultimate/smtinterpol/smtlib2/smtlib.flex
@@ -21,7 +21,7 @@ package de.uni_freiburg.informatik.ultimate.smtinterpol.smtlib2;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import com.github.jhoenicke.javacup.runtime.Symbol;
-import de.uni_freiburg.informatik.ultimate.util.UnifyHash;
+import de.uni_freiburg.informatik.ultimate.util.datastructures.UnifyHash;
 import de.uni_freiburg.informatik.ultimate.smtinterpol.util.MySymbolFactory;
 
 /**
